+++ spec/helpers/codelist_helper_spec.rb
@@ -240,11 +240,11 @@ RSpec.describe CodelistHelper, type: :helper do
       it "returns the BEIS codes and descriptions" do
         options = helper.tags_options
 
-        expect(options.length).to eq 6
+        expect(options.length).to eq 7
         expect(options.first.code).to eq 1
         expect(options.first.description).to eq "Ayrton Fund"
-        expect(options.last.code).to eq 6
-        expect(options.last.description).to eq "Previously reported under GCRF"
+        expect(options.last.code).to eq 7
+        expect(options.last.description).to eq "Previously reported under Newton Fund"
       end
     end
   end